refactor: 항공권 큐레이션 아키텍처 수정
