+++ frontend/react-app/src/pages/home/components/flightSuggestions/curationList/failure/NoFlightCuration.tsx
