+++ frontend/react-app/src/pages/home/components/flightSuggestions/curationList/success/FlightTimeInformation.tsx
